Add a MvcTestSupport for MVC Integration test.
Refactor existing test cases to extend this MvcTestSupport.
