+++ test/DotNetUnknown.Tests/Exception/GlobalExceptionHandlerTests.cs
@@ -1,45 +1,13 @@
 using System.Net;
 using System.Net.Http.Json;
-using System.Reflection;
-using Microsoft.AspNetCore.Hosting;
+using DotNetUnknown.Tests.Support;
 using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.Mvc.Testing;
-using Microsoft.Extensions.DependencyInjection;
 
 namespace DotNetUnknown.Tests.Exception;
 
 [TestFixture]
-internal sealed class GlobalExceptionHandlerTests
+internal sealed class GlobalExceptionHandlerTests : MvcTestSupport
 {
-    [OneTimeSetUp]
-    public void Setup()
-    {
-        _exceptionTestWebAppFactory = new ExceptionTestWebAppFactory();
-        _httpClient = _exceptionTestWebAppFactory.CreateClient();
-    }
-
-    [OneTimeTearDown]
-    public void Teardown()
-    {
-        _exceptionTestWebAppFactory.Dispose();
-        _httpClient.Dispose();
-    }
-
-    private class ExceptionTestWebAppFactory : WebApplicationFactory<Program>
-    {
-        protected override void ConfigureWebHost(IWebHostBuilder builder)
-        {
-            builder.ConfigureServices(services =>
-            {
-                services.AddControllers()
-                    .AddApplicationPart(Assembly.GetExecutingAssembly());
-            });
-        }
-    }
-
-    private WebApplicationFactory<Program> _exceptionTestWebAppFactory;
-    private HttpClient _httpClient;
-
     // Given
     internal static IEnumerable<object[]> ExceptionSourceProvider
     {
@@ -54,7 +22,7 @@ internal sealed class GlobalExceptionHandlerTests
     public async Task TestGlobalExceptionHandler(string url, HttpStatusCode statusCode, string msg)
     {
         // When
-        var httpResponseMessage = await _httpClient.GetAsync(url);
+        var httpResponseMessage = await HttpClient.GetAsync(url);
         // Then
         Assert.That(httpResponseMessage.StatusCode, Is.EqualTo(statusCode));
         var problemDetails = await httpResponseMessage.Content.ReadFromJsonAsync<ProblemDetails>();
